Update @actions packages to their latest versions
This is primarily a fix for GitHub Actions deprecating the `set-state`
command which requires an update of `@actions/core`. To ensure that all
packages are aligned, we update all the `@actions` packages to their
latest versions.

This also switches to Node 16 since that is now the recommended version
and Node 12 is end-of-life.

See: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/
See: https://github.com/actions/typescript-action/commit/87fd0349ba2dff5d5598104cec63562cdec81fa5