Use `pull_request_target` event for triggering the integration test
The integration test is not being triggered for e.g. dependabot PRs.
This changes the event to use `pull_request_target` to ensure that we
have access to creating a status and triggering the workflow.

I'm not 100% about the implications of this, as it will change the
security model of the integration test. However, it should only be
possible to trigger this when you already have write access or when you
have explicitly been approved (since the fork PR workflow permission
is set to "Require approval for all outside collaborators").