Add a Sarif type and validate the SARIF file when reading it
Rather than using a full SARIF spec, the idea here is just to model the bits that
we actually need. This will make the schema must smaller and faster and reduces the
chance of error when the SARIF file slightly deviates from the ectual spec but not
in a way that matters.