Do not check in sourcemaps
Sourcemaps have a lot of changes on every tiny modification of an
input file and increase the size of the repository quite a bit. This
removes the sourcemaps from being checked in.

`npm run build` is deterministic, so we can always generate the
sourcemaps for a specific commit by using the same version of all
dependencies and setting the new
`CODEQL_VARIANT_ANALYSIS_ACTION_GENERATE_SOURCEMAPS` environment
variable.